Continuei a tarefa, mas esse final de semana foi louco e não consegui finalizar os desafios
