+++ semana16/to-do-list/template-express-knex/src/index.ts
@@ -1,13 +1,11 @@
-import express, { Express,  Request, Response } from "express";
+import express, { Express } from "express";
 import knex from "knex";
 import cors from "cors";
 import dotenv from "dotenv";
 import { AddressInfo } from "net";
-import dayjs from "dayjs";
-import { user } from "./types/user";
-import { task } from "./types/task";
-import { getAllUsers, getUserById, insertUser, updateUser } from "./data/UserDao";
-import { getTaskByCreatorUserId, getTaskById, insertTask } from "./data/TaskDao";
+import { getAllUsersEndpoint, getUserByIdEndpoint, getUserByTextEndpoint, insertUserEndpoint, updateUserEndpoint } from "./endpoints/UserService";
+import { getTaskByCreatorUserIdEndpoint, getTaskByIdEndpoint, insertTaskEndpoint } from "./endpoints/TaskService";
+import { insertResponsibleEndpoint } from "./endpoints/ResponsibleUserService";
 
 dotenv.config();
 
@@ -27,150 +25,18 @@ app.use(express.json());
 app.use(cors())
 
 // endpoints aqui
-app.get('/user/all', async (req: Request, res: Response) => {
-  try {
-    const result = await getAllUsers()
 
-    res.status(200).send({users: result})
-  } catch (error) {
-    res.status(400).send(error.message)
-  }
-})
+app.get('/user/all', getAllUsersEndpoint)
+app.get('/user', getUserByTextEndpoint)
+app.get('/user/:id', getUserByIdEndpoint)
+app.post('/user', insertUserEndpoint)
+app.put('/user/edit/:id', updateUserEndpoint)
 
-app.get('/user/:id', async (req: Request, res: Response) => {
-  try {
-    if(!req.params.id) {
-      throw new Error("ID inválido")
-    }
+app.get('/task', getTaskByCreatorUserIdEndpoint)
+app.get('/task/:id', getTaskByIdEndpoint)
+app.post('/task', insertTaskEndpoint)
 
-    const id:string = req.params.id
-    const result = await getUserById(id)
-
-    if(!result) {
-      throw new Error("Usuário não encontrado!")
-    }
-
-    res.status(200).send(result)
-  } catch (error) {
-    res.status(400).send(error.message)
-  }
-})
-
-app.get('/task', async (req: Request, res: Response) => {
-  try {
-    const id: string = req.query.creatorUserId as string
-
-    const result = await getTaskByCreatorUserId(id)
-
-    if(!result) {
-      throw new Error("Usuário não encontrado!")
-    }
-
-    result.map((usertask: any) => {
-      return usertask.limitDate = dayjs(usertask.limitDate).format('DD/MM/YYYY')
-    })
-
-    res.status(200).send({tasks: result})
-  } catch (error) {
-    res.status(400).send(error.message)
-  }
-})
-
-app.get('/task/:id', async (req: Request, res: Response) => {
-  try {
-    if(!req.params.id) {
-      throw new Error("ID inválido")
-    }
-
-    const id:string = req.params.id
-    let result = await getTaskById(id)
-
-    if(!result) {
-      throw new Error("Usuário não encontrado!")
-    }
-
-    result.limitDate = dayjs(result.limitDate).format('DD/MM/YYYY')
-
-    res.status(200).send(result)
-  } catch (error) {
-    res.status(400).send(error.message)
-  }
-})
-
-app.post('/user', async (req: Request, res: Response) => {
-  try {
-    if (!req.body.name || !req.body.nickname || !req.body.email) {
-      throw new Error("Nome, Nickname ou email não informados!")
-    }
-    const {name, nickname, email} = req.body
-    const id: string = String(Date.now())+"-u"
-    const newUser:user = {id: id, name: name, nickname: nickname, email: email}
-    const result = await insertUser(newUser)
-
-    res.status(200).send(result)
-  } catch (error) {
-    res.status(400).send(error.message)
-  }
-})
-
-app.post('/task', async (req: Request, res: Response) => {
-  try {
-    if (!req.body.title || !req.body.description || !req.body.limitDate || !req.body.creatorUserId) {
-      throw new Error("Título, descrição, data limite ou id do criador não informados!")
-    }
-    const {title, description, limitDate, creatorUserId} = req.body
-    const id: string = String(Date.now())+"-t"
-    const limitDateArray = limitDate.split("/")
-    const newLimitDate = limitDateArray[2]+"-"+limitDateArray[1]+"-"+limitDateArray[0]
-    const newTask: task = {id: id, title: title, description: description, limitDate: new Date(newLimitDate), creatorUserId: creatorUserId}
-    const result = await insertTask(newTask)
-
-    res.status(200).send(result)
-  } catch (error) {
-    res.status(400).send(error.message)
-  }
-})
-
-app.put('/user/edit/:id', async (req: Request, res: Response) => {
-  try {
-    const id: string = req.params.id
-    let body = {}
-
-    if (!req.body.name && !req.body.nickname && !req.body.email) {
-      throw new Error("Nenhuma alteração enviada!")
-    }
-
-    if (req.body.name) {
-      if (req.body.name === "") {
-        throw new Error("Novo nome inválido")
-      } else {
-        body = {... body, name: req.body.name}
-      }
-    }
-
-    if (req.body.nickname) {
-      if (req.body.nickname === "") {
-        throw new Error("Novo nome inválido")
-      } else {
-        body = {... body, nickname: req.body.nickname}
-      }
-    }
-
-    if (req.body.email) {
-      if (req.body.email === "") {
-        throw new Error("Novo nome inválido")
-      } else {
-        body = {... body, email: req.body.email}
-      }
-    }
-
-    const result = await updateUser(id, body)
-
-    res.status(200).send(result)
-  } catch (error) {
-    res.status(400).send(error.message)
-  }
-})
+app.post('/task/responsible', insertResponsibleEndpoint)
 
 const server = app.listen(process.env.PORT || 3003, () => {
    if (server) {
