+++ semana16/to-do-list/template-express-knex/src/types/task.ts
@@ -3,5 +3,12 @@ export type task = {
   title: string,
   description: string,
   limitDate: Date,
+  status: status,
   creatorUserId: string
+}
+
+export enum status {
+  to_do = "to-do",
+  doing = "doing",
+  done = "done",
 }
